requirements.txt
# Backend Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Pydantic (V2 Standard)
pydantic>=2.9.0
pydantic-settings>=2.5.0

# LangChain Ecosystem (Wajib Versi 0.3.x biar aman dari Pydantic_V1 error)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.4

# Google Gemini (WAJIB ADA!)
langchain-google-genai>=2.0.0
google-genai>=0.3.0

# Vector Database & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0
transformers>=4.44.0
torch>=2.4.0

# Ingestion Tools
pypdf>=4.3.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
playwright>=1.46.0

# Utilities
python-dotenv>=1.0.1
requests>=2.32.0
httpx>=0.27.0
tenacity>=9.0.0
numpy>=1.26.0